manifest: Add SEMC mogami-common, zeus-common and zeusc
This patch adds the mogami-common and zeus-common repositories
which will be used as a basis for all other devices in order to
avoid the extensive duplication that we have now.

Additionally, it will fetch zeusc which will now derive from
zeus-common and be its own device, instead of being a zeus clone.

All repositories are already in github, just need to be gerrited.

Change-Id: I7036c8444ab9d8854dce183d66e1bbd7fb7eaab2